+++ src/FamilyHubs.SharedKernel.Razor/AddAnother/AddAnotherAutocompleteErrorChecker.cs
@@ -3,10 +3,21 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace FamilyHubs.SharedKernel.Razor.AddAnother;
 
-public record AddAnotherAutocompleteErrorChecker(int? FirstEmptyIndex, int? FirstInvalidNameIndex, int? FirstDuplicateLanguageIndex)
+public class AddAnotherAutocompleteErrorChecker
 {
-    // this would be better as a constructor, but we can't do that until support is added to c#
-    // https://github.com/dotnet/csharplang/issues/7047
+    public IEnumerable<int> EmptyIndexes { get; }
+    public IEnumerable<int> InvalidIndexes { get; }
+    public IEnumerable<IEnumerable<int>> DuplicateIndexes { get; }
+
+    public AddAnotherAutocompleteErrorChecker(
+        IEnumerable<int> emptyIndexes,
+        IEnumerable<int> invalidIndexes,
+        IEnumerable<IEnumerable<int>> duplicateIndexes)
+    {
+        EmptyIndexes = emptyIndexes;
+        InvalidIndexes = invalidIndexes;
+        DuplicateIndexes = duplicateIndexes;
+    }
 
     public static AddAnotherAutocompleteErrorChecker Create(
         IFormCollection form,
@@ -16,56 +27,78 @@ public record AddAnotherAutocompleteErrorChecker(int? FirstEmptyIndex, int? Firs
         //todo: pass the empty value too
         IEnumerable<SelectListItem> validItems)
     {
-        // when js is disables, we won't get the texts, just the values
+        // when js is disabled, we won't get the texts, just the values
         if (form.ContainsKey(textFieldName))
         {
-            // javascript is enabled, we need to work with the texts
-            var texts = form[textFieldName];
-
-            var nameAndIndex = texts
-                .Select((item, index) => new { Item = item, Index = index });
-
-            int? firstEmptyIndex = nameAndIndex.FirstOrDefault(element => element.Item == "")?.Index;
-
-            var validNames = validItems.Select(o => o.Text);
-
-            int? firstInvalidNameIndex =
-                nameAndIndex.FirstOrDefault(x => x.Item != "" && !validNames.Contains(x.Item))?.Index;
-
-            int? firstDuplicateLanguageIndex = nameAndIndex
-                .GroupBy(x => x.Item)
-                .FirstOrDefault(g => g.Key != "" && validNames.Contains(g.Key) && g.Count() > 1)
-                ?.First().Index;
-
-            return new AddAnotherAutocompleteErrorChecker(firstEmptyIndex, firstInvalidNameIndex, firstDuplicateLanguageIndex);
+            return CreateFromJavascriptEnabledPostbackForm(form, textFieldName, validItems);
         }
-        else
+
+        // javascript is disabled, we need to work with the values
+        if (!form.ContainsKey(valuesFieldName))
         {
-            // javascript is disabled, we need to work with the values
-            if (!form.ContainsKey(valuesFieldName))
-            {
-                // we don't have any values, which means we have a single select with no value selected
-                return new AddAnotherAutocompleteErrorChecker(0, null, null);
-            }
-
-            var values = form[valuesFieldName];
-
-            var valuesAndIndex = values
-                .Select((item, index) => new { Item = item, Index = index });
-
-            int? firstEmptyIndex = valuesAndIndex.FirstOrDefault(element => element.Item == "")?.Index;
-
-            int? firstDuplicateLanguageIndex = null;
-            if (values.Count > values.Distinct().Count())
-            {
-                firstDuplicateLanguageIndex =
-                    valuesAndIndex
-                        .GroupBy(x => x.Item)
-                        .FirstOrDefault(g => g.Key != "" && g.Count() > 1)
-                        ?.Skip(1).First().Index;
-            }
-
-            return new AddAnotherAutocompleteErrorChecker(firstEmptyIndex, null, firstDuplicateLanguageIndex);
+            // we don't have any values, which means we have a single select with no value selected
+            return new AddAnotherAutocompleteErrorChecker(
+                new[] {0},
+                Enumerable.Empty<int>(),
+                Enumerable.Empty<IEnumerable<int>>());
         }
+
+        return CreateFromJavascriptDisabledPostbackForm(form, valuesFieldName);
+    }
+
+    private static AddAnotherAutocompleteErrorChecker CreateFromJavascriptEnabledPostbackForm(
+        IFormCollection form,
+        string textFieldName,
+        IEnumerable<SelectListItem> validItems)
+    {
+        // javascript is enabled, we need to work with the texts
+        var texts = form[textFieldName];
+
+        var nameAndIndex = texts
+            .Select((item, index) => new { Item = item, Index = index })
+            .ToArray();
+
+        var emptyIndexes = nameAndIndex
+            .Where(element => element.Item == "")
+            .Select(e => e.Index);
+
+        var validNames = validItems.Select(o => o.Text);
+
+        var invalidIndexes = nameAndIndex
+            .Where(x => x.Item != "" && !validNames.Contains(x.Item))
+            .Select(e => e.Index);
+
+        var duplicateIndexes = nameAndIndex
+            // exclude empty and invalid values from the duplicates check (using emptyIndexes and invalidIndexes, rather than repeat the actual checks)
+            .Where(e => !emptyIndexes.Contains(e.Index) && !invalidIndexes.Contains(e.Index))
+            .GroupBy(x => x.Item)
+            .Where(g => g.Count() > 1)
+            .Select(g => g.Select(x => x.Index));
+
+        return new AddAnotherAutocompleteErrorChecker(emptyIndexes, invalidIndexes, duplicateIndexes);
+    }
+
+    private static AddAnotherAutocompleteErrorChecker CreateFromJavascriptDisabledPostbackForm(
+        IFormCollection form,
+        string valuesFieldName)
+    {
+        var values = form[valuesFieldName];
+
+        var valuesAndIndex = values
+            .Select((item, index) => new { Item = item, Index = index })
+            .ToArray();
+
+        var emptyIndexes = valuesAndIndex
+            .Where(e => e.Item == "")
+            .Select(e => e.Index);
+
+        var duplicateIndexes =
+            valuesAndIndex
+                .Where(e => !emptyIndexes.Contains(e.Index))
+                .GroupBy(x => x.Item)
+                .Where(g => g.Count() > 1)
+                .Select(g => g.Select(x => x.Index));
+
+        return new AddAnotherAutocompleteErrorChecker(emptyIndexes, Enumerable.Empty<int>(), duplicateIndexes);
     }
 }