+++ tests/FamilyHubs.SharedKernel.Razor.UnitTests/AddAnother/AddAnotherAutocompleteErrorCheckerTests.cs
@@ -12,7 +12,10 @@ public class AddAnotherAutocompleteErrorCheckerTests
     public void ShouldBeRoundTripSerializableTest()
     {
         // Arrange
-        var original = new AddAnotherAutocompleteErrorChecker(1, 2, 3);
+        var original = new AddAnotherAutocompleteErrorChecker(
+            Enumerable.Range(1, 3),
+            Enumerable.Range(10, 2),
+            Enumerable.Repeat(Enumerable.Range(1, 3), 2));
 
         // Act
         string serialized = JsonSerializer.Serialize(original);
@@ -20,31 +23,13 @@ public class AddAnotherAutocompleteErrorCheckerTests
 
         // Assert
         Assert.NotNull(deserialized);
-        Assert.Equal(original.FirstDuplicateLanguageIndex, deserialized.FirstDuplicateLanguageIndex);
-        Assert.Equal(original.FirstEmptyIndex, deserialized.FirstEmptyIndex);
-        Assert.Equal(original.FirstInvalidNameIndex, deserialized.FirstInvalidNameIndex);
+        Assert.Equal(original.EmptyIndexes.ToArray(), deserialized.EmptyIndexes.ToArray());
+        Assert.Equal(original.EmptyIndexes.ToArray(), deserialized.EmptyIndexes.ToArray());
+        AssertNestedCollectionsAreEqual(original.DuplicateIndexes, deserialized.DuplicateIndexes);
     }
 
     [Fact]
-    public void ShouldBeRoundTripSerializableTestWithNulls()
-    {
-        // Arrange
-        var original = new AddAnotherAutocompleteErrorChecker(null, null, null);
-
-        // Act
-        string serialized = JsonSerializer.Serialize(original);
-        var deserialized = JsonSerializer.Deserialize<AddAnotherAutocompleteErrorChecker>(serialized);
-
-        // Assert
-        Assert.NotNull(deserialized);
-        Assert.Equal(original.FirstDuplicateLanguageIndex, deserialized.FirstDuplicateLanguageIndex);
-        Assert.Equal(original.FirstEmptyIndex, deserialized.FirstEmptyIndex);
-        Assert.Equal(original.FirstInvalidNameIndex, deserialized.FirstInvalidNameIndex);
-    }
-
-    [Theory]
-    [InlineData(0, null, null)]
-    public void JavascriptDisabled_NoLanguageSelected_ShouldFindFirstEmptyIndex(int? expectedFirstEmptyIndex, int? expectedFirstInvalidNameIndex, int? expectedFirstDuplicateLanguageIndex, params string[] values)
+    public void JavascriptDisabled_NoLanguageSelected_ShouldFindFirstEmptyIndex()
     {
         // Arrange
         var form = new FormCollection(new Dictionary<string, StringValues>());
@@ -54,31 +39,18 @@ public class AddAnotherAutocompleteErrorCheckerTests
             new List<SelectListItem> { new("a", "1"), new("c", "3") });
 
         // Assert
-        Assert.Equal(expectedFirstEmptyIndex, result.FirstEmptyIndex);
-        Assert.Equal(expectedFirstInvalidNameIndex, result.FirstInvalidNameIndex);
-        Assert.Equal(expectedFirstDuplicateLanguageIndex, result.FirstDuplicateLanguageIndex);
+        Assert.Equal(new[] {0}, result.EmptyIndexes);
+        Assert.Empty(result.InvalidIndexes);
+        Assert.Empty(result.InvalidIndexes);
     }
 
     [Theory]
-    [InlineData(0, null, null, "")]
-    [InlineData(0, null, null, "", "b", "c")]
-    [InlineData(1, null, null, "a", "", "c")]
-    [InlineData(1, null, null, "a", "", "c", "")]
-    [InlineData(null, 0, null, "smurf")]
-    [InlineData(null, 1, null, "a", "smurf", "c")]
-    [InlineData(null, null, 0, "a", "a")]
-    [InlineData(null, null, 1, "b", "a", "a", "c")]
-    [InlineData(null, null, 0, "a", "b", "c", "a")]
-    [InlineData(1, 3, null, "b", "", "a", "smurf")]
-    [InlineData(1, 2, null, "b", "", "klingon", "a", "", "smurf")]
-    [InlineData(3, null, 1, "b", "a", "a", "")]
-    [InlineData(1, null, 0, "a", "", "a")]
-    [InlineData(null, 2, 0, "a", "b", "womble", "a")]
-    [InlineData(null, 2, 1, "a", "b", "womble", "b")]
-    [InlineData(2, 3, 1, "a", "b", "", "womble", "b")]
-    [InlineData(2, 3, 0, "a", "b", "", "womble", "b", "a")]
-    [InlineData(0, 6, 1, "", "a", "b", "c", "a", "c", "womble", "b", "a")]
-    public void JavascriptEnabled_ShouldReturnCorrectIndexes(int? expectedFirstEmptyIndex, int? expectedFirstInvalidNameIndex, int? expectedFirstDuplicateLanguageIndex, params string[] texts)
+    [MemberData(nameof(JavascriptEnabled_ShouldReturnCorrectIndexes_TestData))]
+    public void JavascriptEnabled_ShouldReturnCorrectIndexes(
+        IEnumerable<int> expectedEmptyIndexes,
+        IEnumerable<int> expectedInvalidIndexes,
+        IEnumerable<IEnumerable<int>> expectedDuplicateIndexes,
+        string[] texts)
     {
         // Arrange
         var form = new FormCollection(new Dictionary<string, StringValues>
@@ -99,8 +71,45 @@ public class AddAnotherAutocompleteErrorCheckerTests
             });
 
         // Assert
-        Assert.Equal(expectedFirstEmptyIndex, result.FirstEmptyIndex);
-        Assert.Equal(expectedFirstInvalidNameIndex, result.FirstInvalidNameIndex);
-        Assert.Equal(expectedFirstDuplicateLanguageIndex, result.FirstDuplicateLanguageIndex);
+        Assert.Equal(expectedEmptyIndexes.ToArray(), result.EmptyIndexes.ToArray());
+        Assert.Equal(expectedInvalidIndexes.ToArray(), result.InvalidIndexes.ToArray());
+        AssertNestedCollectionsAreEqual(expectedDuplicateIndexes, result.DuplicateIndexes);
+    }
+
+    private void AssertNestedCollectionsAreEqual(
+        IEnumerable<IEnumerable<int>> expected,
+        IEnumerable<IEnumerable<int>> actual)
+    {
+        var expectedList = expected.ToArray();
+        var actualList = actual.ToArray();
+
+        Assert.Equal(expectedList.Length, actualList.Length);
+
+        for (int i = 0; i < expectedList.Length; ++i)
+        {
+            Assert.Equal(expectedList[i], actualList[i]);
+        }
+    }
+
+    public static IEnumerable<object[]> JavascriptEnabled_ShouldReturnCorrectIndexes_TestData()
+    {
+        yield return new object[] { new[] { 0 }, Array.Empty<int>(), Array.Empty<int[]>(), new[] {""} };
+        yield return new object[] { new[] { 0 }, Array.Empty<int>(), Array.Empty<int[]>(), new[] { "", "b", "c" } };
+        yield return new object[] { new[] { 1 }, Array.Empty<int>(), Array.Empty<int[]>(), new[] { "a", "", "c" } };
+        yield return new object[] { new[] { 1, 3 }, Array.Empty<int>(), Array.Empty<int[]>(), new[] { "a", "", "c", "" } };
+        yield return new object[] { Array.Empty<int>(), new[] { 0 }, Array.Empty<int[]>(), new[] { "smurf" } };
+        yield return new object[] { Array.Empty<int>(), new[] { 1 }, Array.Empty<int[]>(), new[] { "a", "smurf", "c" } };
+        yield return new object[] { Array.Empty<int>(), Array.Empty<int>(), new[] { new[] { 0, 1 }}, new[] { "a", "a" } };
+        yield return new object[] { Array.Empty<int>(), Array.Empty<int>(), new[] { new[] { 1, 2 }}, new[] { "b", "a", "a", "c" } };
+        yield return new object[] { Array.Empty<int>(), Array.Empty<int>(), new[] { new[] { 0, 3 }}, new[] { "a", "b", "c", "a" } };
+        yield return new object[] { new[] { 1 }, new[] { 3 }, Array.Empty<int[]>(), new[] { "b", "", "a", "smurf" } };
+        yield return new object[] { new[] { 1, 4 }, new[] { 2, 5 }, Array.Empty<int[]>(), new[] { "b", "", "klingon", "a", "", "smurf" } };
+        yield return new object[] { new[] { 3 }, Array.Empty<int>(), new[] { new[] { 1, 2 }}, new[] { "b", "a", "a", "" } };
+        yield return new object[] { new[] { 1 }, Array.Empty<int>(), new[] { new[] { 0, 2 }}, new[] { "a", "", "a" } };
+        yield return new object[] { Array.Empty<int>(), new[] { 2 }, new[] { new[] { 0, 3 }}, new[] { "a", "b", "womble", "a" } };
+        yield return new object[] { Array.Empty<int>(), new[] { 2 }, new[] { new[] { 1, 3 }}, new[] { "a", "b", "womble", "b" } };
+        yield return new object[] { new[] { 2 }, new[] { 3 }, new[] { new[] { 1, 4 }}, new[] { "a", "b", "", "womble", "b" } };
+        yield return new object[] { new[] { 2 }, new[] { 3 }, new[] { new[] { 0, 5 }, new[] { 1, 4 }}, new[] { "a", "b", "", "womble", "b", "a" } };
+        yield return new object[] { new[] { 0 }, new[] { 6 }, new[] { new[] { 1, 4, 8 }, new[] { 2, 7 }, new[] { 3, 5 } }, new[] { "", "a", "b", "c", "a", "c", "womble", "b", "a" } };
     }
 }